♻️ Move paparazzi files to screenshots package
