+++ coreui/src/test/kotlin/cz/jankotas/screenshottests/coreui/screenshots/PaparazziShowkaseComponentTest.kt
@@ -1,9 +1,10 @@
-package cz.jankotas.screenshottests.coreui
+package cz.jankotas.screenshottests.coreui.screenshots
 
 import androidx.compose.runtime.Composable
 import com.airbnb.android.showkase.models.Showkase
 import com.airbnb.android.showkase.models.ShowkaseBrowserComponent
 import com.android.resources.NightMode
+import cz.jankotas.screenshottests.coreui.getMetadata
 import org.junit.runners.Parameterized
 
 /**
