+++ coreui/src/test/kotlin/cz/jankotas/screenshottests/coreui/screenshots/PaparazziShowkaseTest.kt
@@ -1,4 +1,4 @@
-package cz.jankotas.screenshottests.coreui
+package cz.jankotas.screenshottests.coreui.screenshots
 
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
