+++ coreui/src/test/kotlin/cz/jankotas/screenshottests/coreui/screenshots/ScreenshotPreview.kt
@@ -1,4 +1,4 @@
-package cz.jankotas.screenshottests.coreui
+package cz.jankotas.screenshottests.coreui.screenshots
 
 import androidx.compose.runtime.Composable
